исправил все замечания код ревью
