+++ src/app/blog/page.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import Head from 'next/head';
 import { CardsList } from '@/components/BlockCards/CardsList/CardList';
 import { DescriptionPages } from '@/components/DescriptionPages/DescriptionPages';
 import { blogCards, descriptionPages } from '@/utils/constants';
@@ -7,10 +8,10 @@ import { blogCards, descriptionPages } from '@/utils/constants';
 export default function BlogPage() {
   return (
     <>
-      <head>
+      <Head>
         <title>Трудовые резервы | Блог</title>
         <meta name="title" content="Блог" />
-      </head>
+      </Head>
       <DescriptionPages descriptionPages={descriptionPages} />
       <CardsList blogCards={blogCards} />
     </>
