+++ src/app/coachs/page.tsx
@@ -1,5 +1,6 @@
 'use client';
 
+import Head from 'next/head';
 import { CoachList } from '@/components/CoachCards/CoachList/CoachList';
 import { DescriptionPages } from '@/components/DescriptionPages/DescriptionPages';
 import { coachCards, descriptionPages } from '@/utils/constants';
@@ -7,10 +8,10 @@ import { coachCards, descriptionPages } from '@/utils/constants';
 export default function CoachsPage() {
   return (
     <>
-      <head>
+      <Head>
         <title>Трудовые резервы | Тренерский состав</title>
         <meta name="title" content="Блог" />
-      </head>
+      </Head>
       <DescriptionPages descriptionPages={descriptionPages} />
       <CoachList coachCards={coachCards} />
     </>
