+++ src/hooks/useWindowSize.ts
@@ -1,20 +1,18 @@
 import { useEffect, useState } from 'react';
 
-function useWindowSize() {
-  // const [width, setWidth] = useState(window.innerWidth);
-  const [width, setWidth] = useState({undefined});
-
+function useWindowSize(): number {
+  const [width, setWidth] = useState<number>(0);
   useEffect(() => {
     if (typeof window !== 'undefined') {
-    const handleResize = () => {
-      setWidth(window.innerWidth);
-    };
-  
-    window.addEventListener('resize', handleResize);
-    return () => {
-      window.removeEventListener('resize', handleResize);
-    };
-  }
+      const handleResize = () => {
+        setWidth(window.innerWidth);
+      };
+      window.addEventListener('resize', handleResize);
+      handleResize();
+      return () => {
+        window.removeEventListener('resize', handleResize);
+      };
+    }
   }, []);
 
   return width;
